ingestion/pyproject.toml
[project]
name = "chorus"
version = "0.1.0"
description = "SOP-to-Agent Skill Document platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.0.0",
    "openai>=1.50.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.5.0",
    "jinja2>=3.1.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
